# Symmetrizing read/write wkt