--- a/doc/imports.qbk
+++ b/doc/imports.qbk
@@ -153,6 +153,8 @@
 [import src/examples/io/svg.cpp]
 [import src/examples/io/wkt.cpp]
 [import src/examples/io/read_wkt.cpp]
+[import src/examples/io/to_wkt.cpp]
+[import src/examples/io/from_wkt.cpp]
 
 [import src/examples/strategies/buffer_join_round.cpp]
 [import src/examples/strategies/buffer_join_miter.cpp]