--- a/doc/reference/io/from_wkt.qbk
+++ b/doc/reference/io/from_wkt.qbk
@@ -1,0 +1,23 @@
+[/============================================================================
+  Boost.Geometry (aka GGL, Generic Geometry Library)
+  
+  Copyright (c) 2020 Baidyanath Kundu, Haldia, India.
+
+  Use, modification and distribution is subject to the Boost Software License,
+  Version 1.0. (See accompanying file LICENSE_1_0.txt or copy at
+  http://www.boost.org/LICENSE_1_0.txt)
+=============================================================================/]
+
+[heading Conformance]
+Other libraries refer to this functionality as [*ST_GeomFromText] or [*STGeomFromText].
+That is not done here because Boost.Geometry support more text formats. The name GeomFromText
+is reserved for future usage, which will then have an indication of the used text format.
+
+
+[heading Example]
+[from_wkt]
+
+[heading See also]
+* [link geometry.reference.io.wkt.wkt WKT streaming manipulator]
+* [link geometry.reference.io.wkt.read_wkt Read WKT]
+* [link geometry.reference.io.wkt.to_wkt To WKT]