--- a/doc/reference/io/to_wkt.qbk
+++ b/doc/reference/io/to_wkt.qbk
@@ -2,6 +2,7 @@
   Boost.Geometry (aka GGL, Generic Geometry Library)
 
   Copyright (c) 2009-2014 Barend Gehrels, Amsterdam, the Netherlands.
+  Copyright (c) 2020 Baidyanath Kundu, Haldia, India.
 
   Use, modification and distribution is subject to the Boost Software License,
   Version 1.0. (See accompanying file LICENSE_1_0.txt or copy at