--- a/doc/reference/io/to_wkt.qbk
+++ b/doc/reference/io/to_wkt.qbk
@@ -1,0 +1,24 @@
+[/============================================================================
+  Boost.Geometry (aka GGL, Generic Geometry Library)
+
+  Copyright (c) 2009-2014 Barend Gehrels, Amsterdam, the Netherlands.
+
+  Use, modification and distribution is subject to the Boost Software License,
+  Version 1.0. (See accompanying file LICENSE_1_0.txt or copy at
+  http://www.boost.org/LICENSE_1_0.txt)
+=============================================================================/]
+
+[def __this_function__ to_wkt]
+
+[heading_conformance_ogc __this_function__..AsText]
+[note __this_function__ is not named "AsText" or "as_text" because Boost.Geometry 
+also supports other textformats (svg, dsv)]
+
+[heading Example]
+[to_wkt]
+[to_wkt_output]
+
+[heading See also]
+* [link geometry.reference.io.wkt.read_wkt Read WKT]
+* [link geometry.reference.io.wkt.from_wkt From WKT]
+* [link geometry.reference.io.wkt.wkt WKT streaming manipulator]