--- a/include/boost/geometry/io/wkt/write.hpp
+++ b/include/boost/geometry/io/wkt/write.hpp
@@ -104,34 +104,6 @@ struct output_formatter<std::string>
     }
 };
 
-template <>
-struct output_formatter<std::ostream>
-{
-    template <typename T>
-    static void append(std::ostream& out, T const& val, int significant_digits = -1)
-    {
-        BOOST_STATIC_ASSERT((boost::is_arithmetic<T>::value));
-        if(std::numeric_limits<T>::is_specialized)
-        {
-            if(significant_digits < 0)
-                out.precision(std::numeric_limits<T>::digits10);
-            else
-                out.precision(significant_digits);
-        }
-        out << val;
-    }
-
-    static void append(std::ostream& out, char c, int = -1)
-    {
-        out << c;
-    }
-
-    static void append(std::ostream& out, char const* s, int = -1)
-    {
-        out << s;
-    }
-};
-
 template <typename P, int I, int Count>
 struct stream_coordinate
 {