… doc/html/index.html
… doc/quickref.xml
… doc/src/examples/io/Jamfile
… doc/src/examples/io/from_wkt.cpp
… doc/src/examples/io/to_wkt.cpp
… doc/src/examples/io/wkt.cpp
… test/io/wkt/wkt.cpp