This PR fixes #654 . It adds the following:

- `to_wkt`: This function converts the given geometry to wkt string and returns it instead of modifying a stringstream taken as input like `wkt`.
- `from_wkt`: This function returns the geometry obtained by parsing a wkt string instead of taking in a geometry variable as parameter and storing the geometry in it.
- tests: Tests for both `to_wkt` and `from_wkt` are provided with basically follow the template of already implemented `read_wkt` and `wkt` tests.
- examples and documentation: Example usage of both have been added and documentation has been updated.

This PR is now ready to be reviewed and then merged.