@mloskot 2020-03-05
@sudo-panda FYI, next time you update your PR, then documentation and `doc/src/examples` will be built and verified with GitHub Actions.