@mloskot 2020-03-09
@sudo-panda If you update your PR, you will get #680 change and GitHub Actions will CI check it for you.