@awulkiew 2020-03-20
@barendgehrels

> Both approaches use stringstream internally, so I don't expect significant differences. 

Only for coordinates. For characters like `','` and native strings like `"MULTIPOLYGON("` `std::stringstream` is not used but the characters are directly appended to `std::string`. And nothing stands in the way of detecting native coordinate types and using e.g. C++11 `std::to_string` for them. I proposed to use `std::stringstream` only in case some user had his own user-defined coordinate type and had `operator<<` defined for this coordinate type, then it would be supported out of the box.

Note that the above means that we should probably get rid of `significant_digits` parameter even for `std::string` because some output methods may not support it like e.g. Streams do. E.g. standard functions does not take this parameter as well as `boost::lexical_cast`.

> And adding significant_digits only quite limits it.

This is only an unelegant implementation where this variable is passed for both Strings and Streams but is ignored by the latter. See the default one for stream:
https://github.com/boostorg/geometry/blob/3c62182c82e294d3446173187126ffb0ef6cf513/include/boost/geometry/io/wkt/write.hpp#L73

The more "proper" way would be to e.g. pass `OutputPolicy` into `devarianted_wkt` and this policy would either be empty for Streams or contained whatever data is needed for Strings. But as I mentioned above I think we should consider simply removing this parameter altogether because we may decide to use something different for coordinates in case of Strings.