@mloskot 2020-04-27
(I'm moving the thread https://github.com/boostorg/geometry/pull/670#discussion_r401816305 to main stream of comments as I need more horizontal space :))

### Benchmark update

@sudo-panda I played a bit more with benchmarking to better understand where your implementation sits versus the stream manipulator `wkt` and `to_wkt` as the basic wrapper.

First, I defined `to_wkt` as the basic wrapper according to what we discussed with @barendgehrels in https://github.com/boostorg/geometry/pull/670#discussion_r394707569

```cpp
template <typename Geometry>
inline std::string to_wkt_wrapper(Geometry const& geometry)
{
    std::stringstream ss;
    ss << boost::geometry::wkt(geometry);
    return ss.str();
}
template <typename Geometry>
inline std::string to_wkt_wrapper(Geometry const& geometry, int significant_digits)
{
    std::stringstream ss;
    ss.precision(significant_digits);
    ss << boost::geometry::wkt(geometry);
    return ss.str();
}
```

then I added that to my [benchmark_boost_geometry_wkt.cpp](https://github.com/mloskot/string_benchmark/blob/master/benchmark_boost_geometry_wkt.cpp) as `to_wkt_wrap`.

Here is the result:

```
|     Group      |   Experiment    |   Prob. Space   |     Samples     |   Iterations    |    Baseline     |  us/Iteration   | Iterations/sec  |
|:--------------:|:---------------:|:---------------:|:---------------:|:---------------:|:---------------:|:---------------:|:---------------:|
|wkt             | to_string       |               1 |              30 |          262144 |         1.00000 |         0.90276 |      1107714.67 |
|wkt             | to_string       |             256 |              30 |              16 |         1.00000 |       245.25000 |         4077.47 |
|wkt             | to_string       |             512 |              30 |               8 |         1.00000 |       490.37500 |         2039.26 |
|wkt             | to_string       |            1024 |              30 |               4 |         1.00000 |       982.00000 |         1018.33 |
|wkt             | stringstream    |               1 |              30 |          131072 |         1.66540 |         1.50346 |       665134.15 |
|wkt             | stringstream    |             256 |              30 |              16 |         1.26988 |       311.43750 |         3210.92 |
|wkt             | stringstream    |             512 |              30 |               8 |         1.24369 |       609.87500 |         1639.68 |
|wkt             | stringstream    |            1024 |              30 |               4 |         1.25305 |      1230.50000 |          812.68 |
|wkt             | lexical_cast    |               1 |              30 |          131072 |         1.76430 |         1.59274 |       627847.71 |
|wkt             | lexical_cast    |             256 |              30 |              16 |         1.40979 |       345.75000 |         2892.26 |
|wkt             | lexical_cast    |             512 |              30 |               8 |         1.40836 |       690.62500 |         1447.96 |
|wkt             | lexical_cast    |            1024 |              30 |               4 |         1.40860 |      1383.25000 |          722.94 |
|wkt             | wkt             |               1 |              30 |          131072 |         1.99188 |         1.79819 |       556115.61 |
|wkt             | wkt             |             256 |              30 |              16 |         1.49847 |       367.50000 |         2721.09 |
|wkt             | wkt             |             512 |              30 |               8 |         1.41601 |       694.37500 |         1440.14 |
|wkt             | wkt             |            1024 |              30 |               4 |         1.40657 |      1381.25000 |          723.98 |
|wkt             | to_wkt_wrap     |               1 |              30 |          131072 |         1.96920 |         1.77772 |       562519.04 |
|wkt             | to_wkt_wrap     |             256 |              30 |              16 |         1.42635 |       349.81250 |         2858.67 |
|wkt             | to_wkt_wrap     |             512 |              30 |               8 |         1.43946 |       705.87500 |         1416.68 |
|wkt             | to_wkt_wrap     |            1024 |              30 |               4 |         1.39282 |      1367.75000 |          731.13 |
|wkt             | to_wkt_wrap_3di |               1 |              30 |          131072 |         1.90817 |         1.72262 |       580511.72 |
|wkt             | to_wkt_wrap_3di |             256 |              30 |              16 |         1.33384 |       327.12500 |         3056.94 |
|wkt             | to_wkt_wrap_3di |             512 |              30 |               8 |         1.32858 |       651.50000 |         1534.92 |
|wkt             | to_wkt_wrap_3di |            1024 |              30 |               4 |         1.32612 |      1302.25000 |          767.90 |
|wkt             | to_wkt_wrap_9di |               1 |              30 |          131072 |         3.01452 |         2.72139 |       367459.31 |
|wkt             | to_wkt_wrap_9di |             256 |              30 |              16 |         1.56600 |       384.06250 |         2603.74 |
|wkt             | to_wkt_wrap_9di |             512 |              30 |               8 |         1.53887 |       754.62500 |         1325.16 |
|wkt             | to_wkt_wrap_9di |            1024 |              30 |               4 |         1.56263 |      1534.50000 |          651.68 |
|wkt             | to_wkt          |               1 |              30 |           65536 |         2.93074 |         2.64575 |       377964.38 |
|wkt             | to_wkt          |             256 |              30 |              16 |         2.43552 |       597.31250 |         1674.17 |
|wkt             | to_wkt          |             512 |              30 |               8 |         2.44099 |      1197.00000 |          835.42 |
|wkt             | to_wkt          |            1024 |              30 |               4 |         2.41802 |      2374.50000 |          421.14 |
|wkt             | to_wkt_3dig     |               1 |              30 |          131072 |         1.99573 |         1.80167 |       555041.75 |
|wkt             | to_wkt_3dig     |             256 |              30 |              16 |         1.84174 |       451.68750 |         2213.92 |
|wkt             | to_wkt_3dig     |             512 |              30 |               8 |         1.83533 |       900.00000 |         1111.11 |
|wkt             | to_wkt_3dig     |            1024 |              30 |               4 |         1.83656 |      1803.50000 |          554.48 |
|wkt             | to_wkt_9dig     |               1 |              30 |          131072 |         2.87525 |         2.59566 |       385258.86 |
|wkt             | to_wkt_9dig     |             256 |              30 |              16 |         2.00739 |       492.31250 |         2031.23 |
|wkt             | to_wkt_9dig     |             512 |              30 |               8 |         2.00408 |       982.75000 |         1017.55 |
|wkt             | to_wkt_9dig     |            1024 |              30 |               4 |         2.00178 |      1965.75000 |          508.71 |
```

Next, I tried a basic optimisation: do not use `std::stringstream` if (theoretically) not necessary. Here is my version of `output_formatter<std::string>::append<T>`:

```cpp
template <typename T>
static void append(std::string& out, T const& val, int significant_digits = -1)
{
    BOOST_STATIC_ASSERT((boost::is_arithmetic<T>::value));
    if (significant_digits >= 0)
    {
        std::stringstream ss;
        ss.precision(significant_digits);
        ss << val;
        out += ss.str();
    }
    else
    {
        out += std::to_string(val);
    }
}
```

Now, look at how this speeds up your `to_wkt` for the case of the default number of digits against the `to_wkt_wrap` from the results above as well as against `to_wkt_3dig` and `to_wkt_9dig` which go via `std::stringstream`:

```
|wkt             | to_wkt          |               1 |              30 |          131072 |         1.77079 |         1.64599 |       607537.67 |
|wkt             | to_wkt          |             256 |              30 |              16 |         1.04082 |       255.00000 |         3921.57 |
|wkt             | to_wkt          |             512 |              30 |               8 |         1.03724 |       508.25000 |         1967.54 |
|wkt             | to_wkt          |            1024 |              30 |               4 |         1.03931 |      1018.00000 |          982.32 |
|wkt             | to_wkt_3dig     |               1 |              30 |          131072 |         1.94660 |         1.80941 |       552666.31 |
|wkt             | to_wkt_3dig     |             256 |              30 |              16 |         1.85459 |       454.37500 |         2200.83 |
|wkt             | to_wkt_3dig     |             512 |              30 |               8 |         1.92270 |       942.12500 |         1061.43 |
|wkt             | to_wkt_3dig     |            1024 |              30 |               4 |         1.91756 |      1878.25000 |          532.41 |
|wkt             | to_wkt_9dig     |               1 |              30 |          131072 |         2.82303 |         2.62407 |       381087.51 |
|wkt             | to_wkt_9dig     |             256 |              30 |              16 |         2.02883 |       497.06250 |         2011.82 |
|wkt             | to_wkt_9dig     |             512 |              30 |               8 |         2.01913 |       989.37500 |         1010.74 |
|wkt             | to_wkt_9dig     |            1024 |              30 |               4 |         2.01174 |      1970.50000 |          507.49 |
```

### Observations

1. Your current `to_wkt` makes lots of calls to `std::stringstream::str()`, while the `to_wkt_wrap` makes single call to `str()` at the end, that makes the former much slower.
2. The modified `to_wkt_` using `to_string` is fast becasue `to_string` calls `sprintf` (or equivalent) and not `strignstream`.

Regarding the 2., on MSVC++, it calls `scprintf` to get the required length, then constructs `string s(length, '\0')` as buffer, then `sprintf(&s[0], len + 1, "%f", value)` to convert `value` into string buffer `s`.

### Conclusion

If `output_formatter<std::string>` is implemented in terms of `sprintf` it will offer best performance which is also ~50% better performance than `to_wkt_wrap` implemented as thin wrapper on top of the `wkt` stream manipulator.

Folks, please, let me know if my reasoning is correct or not.